ufiber_create: set eflags properly, plug leak
eflags was not being initialized.  ufiber_create() is now split into two
parts: the C part, which is relatively architecture-independent, and the
asm part which is not.  The __ufiber_create() function (asm) takes a
pointer to the return address and sets up the stack for the initial
context switch.

Also, a rather large memory leak is fixed, in which stacks were being
reallocated instead of reused.