Push all stack handling code into arch.S
_ufiber_create is now passed a pointer to the bottom of the stack, as
well as the stack size.  The assembly routine can then add stack_size to
the stack pointer, or not, depending on the direction of stack growth.

This makes the C code more portable, at the cost of some additional
complexity in the assembly routines.  This is as it should be--porting
ufibers should not involve writing any C code at all.